@@ -1,6 +1,6 @@
 {
   "name": "wasabi-common",
-  "version": "2.1.0",
+  "version": "2.1.1",
   "description": "lib/index.js",
   "main": "lib/index.js",
   "typings": "lib/index",
@@ -13,20 +13,21 @@
   "author": "kamilbukum@gmail.com",
   "license": "MIT",
   "devDependencies": {
-    "@types/chai": "^4.0.3",
-    "@types/mocha": "^2.2.41",
-    "@types/sinon": "^2.3.3",
-    "chai": "^4.1.1",
+    "@types/chai": "^4.0.5",
+    "@types/mocha": "^2.2.44",
+    "@types/sinon": "^4.0.0",
+    "chai": "^4.1.2",
     "colors": "^1.1.2",
-    "electron": "^1.7.5",
-    "electron-mocha": "^4.0.0",
-    "mocha": "^3.5.0",
-    "rimraf": "^2.6.1",
+    "electron": "^1.7.9",
+    "electron-mocha": "^5.0.0",
+    "mocha": "^4.0.1",
+    "rimraf": "^2.6.2",
     "shx": "^0.2.2",
-    "sinon": "^3.2.0",
+    "sinon": "^4.1.2",
     "ts-node": "^3.3.0",
-    "typedoc": "^0.8.0",
-    "typescript": "^2.4.2"
+    "tslint": "^5.8.0",
+    "typedoc": "^0.9.0",
+    "typescript": "^2.6.1"
   },
   "repository": {
     "type": "git",
