@@ -1,13 +1,14 @@
 import {expect} from "chai";
 import Set from "wasabi-common/lib/collection/Set";
 
+/* tslint:disable no-unused-expression */
 describe("collection/Map", () => {
     it("constructor", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set).to.be.exist;
     });
     it("contains", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.contains("element")).to.be.false;
         set.add("element");
         expect(set.contains("element")).to.be.true;
@@ -15,7 +16,7 @@ describe("collection/Map", () => {
         expect(set.contains("element")).to.be.false;
     });
     it("indexOf", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.indexOf("element")).to.be.eq(-1);
         set.add("element");
         expect(set.indexOf("element")).to.be.be.eq(0);
@@ -23,7 +24,7 @@ describe("collection/Map", () => {
         expect(set.indexOf("element")).to.be.eq(-1);
     });
     it("get", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.get(0)).to.be.undefined;
         set.add("element");
         expect(set.get(0)).to.be.be.eq("element");
@@ -31,7 +32,7 @@ describe("collection/Map", () => {
         expect(set.get(0)).to.be.undefined;
     });
     it("add", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.get(0)).to.be.undefined;
         set.add("element");
         expect(set.get(0)).to.be.be.eq("element");
@@ -39,18 +40,18 @@ describe("collection/Map", () => {
         expect(set.get(0)).to.be.undefined;
     });
     it("addAll", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.get(0)).to.be.undefined;
         set.addAll(["element"]);
         expect(set.get(0)).to.be.be.eq("element");
         set.addAll(set);
         expect(set.get(0)).to.be.be.eq("element");
-        let anotherSet = new Set(["test"]);
+        const anotherSet = new Set(["test"]);
         set.addAll(anotherSet);
         expect(set.get(1)).to.be.be.eq("test");
     });
     it("remove", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.get(0)).to.be.undefined;
         set.add("element");
         expect(set.get(0)).to.be.be.eq("element");
@@ -58,7 +59,7 @@ describe("collection/Map", () => {
         expect(set.get(0)).to.be.undefined;
     });
     it("length", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.length).to.be.eq(0);
         set.add("element");
         expect(set.length).to.be.eq(1);
@@ -66,7 +67,7 @@ describe("collection/Map", () => {
         expect(set.length).to.be.eq(0);
     });
     it("addArray", () => {
-        let set = new Set([]);
+        const set = new Set([]);
         expect(set.get(0)).to.be.undefined;
         Set.addArray(["element"], set);
         expect(set.get(0)).to.be.be.eq("element");
@@ -75,12 +76,12 @@ describe("collection/Map", () => {
     });
 
     it("addSet", () => {
-        let set = new Set(["element"]);
+        const set = new Set(["element"]);
         expect(set.get(0)).to.be.be.eq("element");
         Set.addSet(set, set);
         expect(set.get(0)).to.be.be.eq("element");
-        let anotherSet = new Set(["test"]);
+        const anotherSet = new Set(["test"]);
         Set.addSet(anotherSet, set);
         expect(set.get(1)).to.be.be.eq("test");
-    })
+    });
 });
