@@ -1,6 +1,7 @@
 import {expect} from "chai";
 import {default as Tree} from "wasabi-common/lib/collection/Tree";
 
+/* tslint:disable no-unused-expression */
 describe("collection/Tree", () => {
     it("constructor", () => {
         let map = new Tree();
@@ -9,11 +10,11 @@ describe("collection/Tree", () => {
         map = new Tree({});
         expect(map.get()).to.be.deep.eq({});
 
-        let obj = {
+        const obj = {
             key1: {
-                childOfKey1: ""
+                childOfKey1: "",
             },
-            key2: ""
+            key2: "",
         };
 
         map = new Tree(obj);
@@ -21,11 +22,11 @@ describe("collection/Tree", () => {
     });
 
     it("put & get", () => {
-        let expectedValue = "test";
-        let map = new Tree();
+        const expectedValue = "test";
+        const map = new Tree();
         map.put("key1", true);
         expect(map.get()).to.be.deep.eq({
-            key1: true
+            key1: true,
         });
         expect(map.get("key1")).to.be.deep.eq(true);
 
@@ -34,9 +35,9 @@ describe("collection/Tree", () => {
         expect(map.get()).to.be.deep.eq({
             key1: {
                 key2: {
-                    key3: expectedValue
-                }
-            }
+                    key3: expectedValue,
+                },
+            },
         });
 
         expect(map.get("key1.key2.key3")).to.be.deep.eq(expectedValue);
@@ -47,83 +48,83 @@ describe("collection/Tree", () => {
         expect(map.get("key5")).to.be.undefined;
 
         map.put("key3", {
-            "key4": "example"
+            key4: "example",
         });
 
         expect(map.get("key3", "key4")).to.be.eq("example");
     });
 
     it("putAll", () => {
-        let map = new Tree();
+        const map = new Tree();
         map.putAll({
-            "key1": true
+            key1: true,
         });
         expect(map.get()).to.be.deep.eq({
-            key1: true
+            key1: true,
         });
         expect(map.get("key1")).to.be.deep.eq(true);
 
-        let obj = {
+        const obj = {
             key1: {
                 key2: {
-                    key3: "example"
-                }
-            }
+                    key3: "example",
+                },
+            },
         };
         map.putAll(obj);
         expect(map.get()).to.be.deep.eq(obj);
         expect(map.get("key1.key2.key3")).to.be.deep.eq(obj.key1.key2.key3);
         expect(map.get("key1", "key2.key3")).to.be.deep.eq(obj.key1.key2.key3);
-    })
+    });
 
     it("tree", () => {
-        let map = new Tree();
-        let obj = {
+        const map = new Tree();
+        const obj = {
             key1: {
                 key2: {
-                    key3: "example"
-                }
-            }
+                    key3: "example",
+                },
+            },
         };
         map.putAll(obj);
         expect(map.tree("key1")).to.be.deep.eq(new Tree(obj.key1));
         expect(map.tree("key1.key2")).to.be.deep.eq(new Tree(obj.key1.key2));
-    })
+    });
 
     it("remove", () => {
-        let map = new Tree();
-        let obj = {
+        const map = new Tree();
+        const obj = {
             key1: {
                 key2: {
-                    key3: "example"
-                }
-            }
+                    key3: "example",
+                },
+            },
         };
         map.putAll(obj);
         expect(map.tree("key1")).to.be.deep.eq(new Tree(obj.key1));
         expect(map.remove("key1.key2")).to.be.true;
         expect(map.get()).to.be.deep.eq({
-            key1: {}
+            key1: {},
         });
         expect(map.remove("key1.key2")).to.be.false;
         expect(map.remove("key1", "key2")).to.be.false;
         expect(map.remove("a")).to.be.false;
         expect(map.remove("key1")).to.be.true;
         expect(map.remove()).to.be.false;
-    })
+    });
 
     it("clear", () => {
-        let map = new Tree<string>();
-        let obj = {
+        const map = new Tree<string>();
+        const obj = {
             key1: {
                 key2: {
-                    key3: "example"
-                }
-            }
+                    key3: "example",
+                },
+            },
         };
         map.putAll(obj);
         expect(map.tree("key1")).to.be.deep.eq(new Tree(obj.key1));
         map.clear();
-        expect(map.get()).to.be.deep.eq({})
-    })
+        expect(map.get()).to.be.deep.eq({});
+    });
 });
