@@ -2,6 +2,7 @@ import {expect} from "chai";
 import Binder from "wasabi-common/lib/lang/Binder";
 import {ExampleSubClass} from "./BinderSample";
 
+/* tslint:disable no-unused-expression */
 describe("lang/Binder", () => {
     it("constructor", () => {
         class Sample extends Binder {
@@ -10,13 +11,13 @@ describe("lang/Binder", () => {
             }
         }
 
-        let sample = new Sample();
+        const sample = new Sample();
         expect(sample.test()).to.be.eq(sample);
 
     });
 
     it("bind", () => {
-        let example = new ExampleSubClass();
+        const example = new ExampleSubClass();
         Binder.bind(example, "test1");
         expect(example.test1()).to.be.undefined;
 
@@ -35,12 +36,12 @@ describe("lang/Binder", () => {
         expect(example.test1()).to.be.eq("arrow");
     });
     it("bindAll", () => {
-        let example = new ExampleSubClass();
+        const example = new ExampleSubClass();
         Binder.bindAll(example);
         expect(example.test1()).to.be.undefined;
         expect(example.test2()).to.be.undefined;
         expect(example.test5()).to.be.undefined;
         Binder.bindAll(example, "test1");
         expect(example.test1()).to.be.undefined;
-    })
+    });
 });
