@@ -1,6 +1,7 @@
-import Chars from "wasabi-common/lib/types/Chars"
 import {expect} from "chai";
+import Chars from "wasabi-common/lib/types/Chars";
 
+/* tslint:disable no-unused-expression */
 describe("util/Chars", () => {
     it("isNumber", () => {
         expect(Chars.isNumber("3".charCodeAt(0))).to.be.true;
