@@ -1,7 +1,8 @@
+import {expect} from "chai";
 import Objects from "wasabi-common/lib/types/Objects";
 import Types from "wasabi-common/lib/util/Types";
-import {expect} from "chai";
 
+/* tslint:disable no-unused-expression */
 describe("util/Objects", () => {
     it("has", () => {
         expect(Objects.has({})).to.be.false;
@@ -13,21 +14,21 @@ describe("util/Objects", () => {
     });
 
     it("getLength", () => {
-        let object = {
+        const object = {
             a: "4",
             b: 5,
-            c: [6, 7]
+            c: [6, 7],
         };
         expect(Objects.getLength(object)).to.deep.eq(3);
     });
 
     it("remove", () => {
-        let actual = {
+        const actual = {
             a: "example",
-            b: "example2"
+            b: "example2",
         };
-        let expected = {
-            b: "example2"
+        const expected = {
+            b: "example2",
         };
         expect(Objects.remove(actual, "a")).to.be.deep.eq(expected);
         Objects.remove(actual, "a");
@@ -36,12 +37,12 @@ describe("util/Objects", () => {
     });
 
     it("removeValue", () => {
-        let actual = {
+        const actual = {
             a: "example",
-            b: "example2"
+            b: "example2",
         };
-        let expected = {
-            b: "example2"
+        const expected = {
+            b: "example2",
         };
         expect(Objects.removeValue(actual, "example")).to.be.deep.eq(expected);
         Objects.removeValue(actual, "example");
@@ -49,26 +50,26 @@ describe("util/Objects", () => {
     });
 
     it("map", () => {
-        let object = {
+        const object = {
             a: "4",
             b: 5,
-            c: [6, 7]
+            c: [6, 7],
         };
-        let expectedArray = ["String", "Number", "Array"];
-        let array: any[] = Objects.map(object, (item) => {
+        const expectedArray = ["String", "Number", "Array"];
+        const array: any[] = Objects.map(object, (item) => {
             return Types.getName(item);
         });
         expect(array).to.deep.eq(expectedArray);
     });
 
     it("forEach", () => {
-        let object = {
+        const object = {
             a: "4",
             b: 5,
-            c: [6, 7]
+            c: [6, 7],
         };
-        let expectedArray = ["String", "Number", "Array"];
-        let array: any[] = [];
+        const expectedArray = ["String", "Number", "Array"];
+        const array: any[] = [];
         Objects.forEach(object, (item) => {
             array.push(Types.getName(item));
         });
@@ -76,29 +77,29 @@ describe("util/Objects", () => {
     });
 
     it("getKeys", () => {
-        let object = {
+        const object = {
             a: "4",
             b: 5,
-            c: [6, 7]
+            c: [6, 7],
         };
         expect(Objects.getKeys(object)).to.deep.eq(["a", "b", "c"]);
     });
 
     it("addValue", () => {
-        let object = {
+        const object = {
             a: "4",
             b: 5,
-            c: [6, 7]
+            c: [6, 7],
         };
         Objects.addValue(object, "d", "new element");
         let expected: any = {
             a: "4",
             b: 5,
             c: [6, 7],
-            d: "new element"
+            d: "new element",
         };
         expect(object).to.deep.eq(expected);
-        let keys = ["f", "g"];
+        const keys = ["f", "g"];
         Objects.addValue(object, "e", "Example", keys);
         expected = {
             a: "4",
@@ -107,91 +108,91 @@ describe("util/Objects", () => {
             d: "new element",
             e: {
                 f: {
-                    g: "Example"
-                }
-            }
+                    g: "Example",
+                },
+            },
         };
         expect(object).to.deep.eq(expected);
     });
 
     it("getValue", () => {
-        let object = {
+        const object = {
             a: "4",
             b: 5,
             c: [6, 7],
             d: "new element",
             e: {
                 f: {
-                    g: "Example"
-                }
-            }
+                    g: "Example",
+                },
+            },
         };
         expect(Objects.getValue(object, "e", ["f", "g"])).to.deep.eq("Example");
     });
 
     it("clone", () => {
-        let obj = {};
-        let objectClone = Objects.clone(obj, []);
+        const obj = {};
+        const objectClone = Objects.clone(obj, []);
         expect(obj).to.deep.eq(objectClone);
     });
 
     it("merge", () => {
-        let obj = {
+        const obj = {
             a: 5,
-            c: {
-                d: "4"
-            },
-            b: "Deneme"
-        };
-
-        let obj2 = {
             b: "Deneme",
-            c: {
-                e: ["Example", "Ex"]
-            },
-            d: "Deneme"
-        };
-        let mergedObject = Objects.merge(obj, obj2);
-
-        let expectedObj = {
-            a: 5,
             c: {
                 d: "4",
-                e: ["Example", "Ex"]
             },
+        };
+
+        const obj2 = {
             b: "Deneme",
-            d: "Deneme"
+            c: {
+                e: ["Example", "Ex"],
+            },
+            d: "Deneme",
+        };
+        const mergedObject = Objects.merge(obj, obj2);
+
+        const expectedObj = {
+            a: 5,
+            b: "Deneme",
+            c: {
+                d: "4",
+                e: ["Example", "Ex"],
+            },
+            d: "Deneme",
         };
         expect(mergedObject).to.deep.eq(expectedObj);
     });
 
     it("mergeDefaults", () => {
-        let defaults = {
+        const defaults = {
             a: 5,
+            b: "defaultDeneme",
             c: {
                 d: "4",
-                e: ["Default"]
+                e: ["Default"],
             },
-            b: "defaultDeneme"
         };
 
-        let obj2 = {
+        const obj2 = {
             b: "obj2 deneme",
             c: {
-                e: ["Example", "Ex"]
+                e: ["Example", "Ex"],
             },
-            d: "Deneme"
+            d: "Deneme",
         };
-        let mergedObject = Objects.mergeDefaults(defaults, obj2);
+        const mergedObject = Objects.mergeDefaults(defaults, obj2);
 
-        let expectedObj = {
+        const expectedObj = {
             a: 5,
+            b: "obj2 deneme",
             c: {
                 d: "4",
-                e: ["Example", "Ex"]
+                e: ["Example", "Ex"],
             },
-            b: "obj2 deneme",
-            d: "Deneme"
+            d: "Deneme",
         };
         expect(mergedObject).to.deep.eq(expectedObj);
     });
