@@ -1,6 +1,7 @@
-import Validations from "wasabi-common/lib/util/Validations"
 import {expect} from "chai";
+import Validations from "wasabi-common/lib/util/Validations";
 
+/* tslint:disable no-unused-expression */
 describe("util/Validations", () => {
     it("isPrimitive", () => {
         expect(Validations.isPrimitive("")).to.be.true;
@@ -13,7 +14,8 @@ describe("util/Validations", () => {
         expect(Validations.isPrimitive(new Date())).to.be.false;
         expect(Validations.isPrimitive([])).to.be.false;
         expect(Validations.isPrimitive({})).to.be.false;
-        expect(Validations.isPrimitive(function () {
+        expect(Validations.isPrimitive(() => {
+            /* */
         })).to.be.false;
         expect(Validations.isPrimitive(new RegExp(".*"))).to.be.false;
     });
@@ -29,7 +31,8 @@ describe("util/Validations", () => {
         expect(Validations.isJsonType(new Date())).to.be.false;
         expect(Validations.isJsonType([])).to.be.true;
         expect(Validations.isJsonType({})).to.be.true;
-        expect(Validations.isJsonType(function () {
+        expect(Validations.isJsonType(() => {
+            /* */
         })).to.be.false;
         expect(Validations.isJsonType(new RegExp(".*"))).to.be.false;
     });
@@ -46,7 +49,8 @@ describe("util/Validations", () => {
         expect(Validations.isNativeType(new RegExp(".*"))).to.be.true;
         expect(Validations.isNativeType([])).to.be.true;
         expect(Validations.isNativeType({})).to.be.true;
-        expect(Validations.isNativeType(function () {
+        expect(Validations.isNativeType(() => {
+            /* */
         })).to.be.true;
         expect(Validations.isNativeType(new RegExp(".*"))).to.be.true;
     });
@@ -158,8 +162,8 @@ describe("util/Validations", () => {
     });
 
     it("isFunction", () => {
-        expect(Validations.isFunction(function () {
-
+        expect(Validations.isFunction(() => {
+            /* */
         })).to.be.true;
         expect(Validations.isFunction("")).to.be.false;
         expect(Validations.isFunction("Danina")).to.be.false;
@@ -187,7 +191,6 @@ describe("util/Validations", () => {
         expect(Validations.isUndefined([])).to.be.false;
         expect(Validations.isUndefined({})).to.be.false;
     });
-
 
     it("has", () => {
         expect(Validations.has(null)).to.be.false;
