@@ -1,7 +1,7 @@
 export interface MapItems<V> {
-    [key: string]: V
+    [key: string]: V;
 
-    [key: number]: V
+    [key: number]: V;
 }
 
 export default class Collection {
@@ -11,9 +11,9 @@ export default class Collection {
      * @param {(value: T, key: K) => U} callback
      * @returns {Array<U>}
      */
-    public static map<T, U, K>(list: MapItems<T> | Array<T>, callback: (value: T, key: K) => U): Array<U> {
+    public static map<T, U, K>(list: MapItems<T> | T[], callback: (value: T, key: K) => U): U[] {
         if (Object.prototype.toString.call(list) === "[object Array]") {
-            return this.mapArray(list as Array<T>, callback);
+            return this.mapArray(list as T[], callback);
         }
         return this.mapObject(list as MapItems<T>, callback);
     }
@@ -24,13 +24,13 @@ export default class Collection {
      * @param {(value: T, key: K) => U} callback
      * @returns {Array<U>}
      */
-    public static mapObject<T, U, K>(map: MapItems<T>, callback: (value: T, key: K) => U): Array<U> {
-        let items: Array<U> = [];
+    public static mapObject<T, U, K>(map: MapItems<T>, callback: (value: T, key: K) => U): U[] {
+        const items: U[] = [];
 
-        for (let key in map) {
+        for (const key in map) {
             if (map.hasOwnProperty(key)) {
-                let result = callback(map[key], key as any);
-                if (result != undefined) {
+                const result = callback(map[key], key as any);
+                if (result !== undefined) {
                     items[items.length] = result as any;
                 }
             }
@@ -44,12 +44,14 @@ export default class Collection {
      * @param {(value: T, key: K) => U} callback
      * @returns {Array<U>}
      */
-    public static mapArray<T, U, K>(array: Array<T>, callback: (value: T, key: K) => U): Array<U> {
-        let items: Array<U> = [];
-        if (!array) return items;
+    public static mapArray<T, U, K>(array: T[], callback: (value: T, key: K) => U): U[] {
+        const items: U[] = [];
+        if (!array) {
+            return items;
+        }
         for (let i = 0; i < array.length; i++) {
-            let result = callback(array[i], i as any);
-            if (result != undefined) {
+            const result = callback(array[i], i as any);
+            if (result !== undefined) {
                 items[items.length] = result as any;
             }
         }
@@ -62,9 +64,9 @@ export default class Collection {
      * @param {(value: T, key: K) => (void | boolean)} callback
      * @returns {boolean}
      */
-    public static forEach<T, U, K>(list: MapItems<T> | Array<T>, callback: (value: T, key: K) => void | boolean): boolean {
+    public static forEach<T, U, K>(list: MapItems<T> | T[], callback: (value: T, key: K) => void | boolean): boolean {
         if (Object.prototype.toString.call(list) === "[object Array]") {
-            return this.forEachArray(list as Array<T>, callback);
+            return this.forEachArray(list as T[], callback);
         }
         return this.forEachObject(list as MapItems<T>, callback);
     }
@@ -76,7 +78,7 @@ export default class Collection {
      * @returns {Array<U>}
      */
     public static forEachObject<T, U, K>(map: MapItems<T>, callback: (value: T, key: K) => boolean | any): boolean {
-        for (let key in map) {
+        for (const key in map) {
             if (map.hasOwnProperty(key)) {
                 if (callback(map[key], key as any) === false) {
                     return false;
@@ -92,8 +94,10 @@ export default class Collection {
      * @param {(value: T, key: K) => (void | boolean)} callback
      * @returns {boolean}
      */
-    public static forEachArray<T, U, K>(array: Array<T>, callback: (value: T, key: K) => void | boolean): boolean {
-        if (!array) return true;
+    public static forEachArray<T, U, K>(array: T[], callback: (value: T, key: K) => void | boolean): boolean {
+        if (!array) {
+            return true;
+        }
         for (let i = 0; i < array.length; i++) {
             if (callback(array[i], i as any) === false) {
                 return false;
