@@ -1,18 +1,11 @@
 export default class Binder {
     /**
      *
-     */
-    public constructor() {
-        Binder.bindAll(this);
-    }
-
-    /**
-     *
      * @param instance
      * @param {string} key
      * @param {Function} fn
      */
-    public static bind(instance: any, key: string, fn?: Function) {
+    public static bind(instance: any, key: string, /* tslint:disable */fn?: Function) {
         if (!fn) {
             fn = instance[key];
         }
@@ -26,13 +19,23 @@ export default class Binder {
      */
     public static bindAll(instance: any, ...keys: string[]) {
         if (keys.length === 0) {
-            keys = Object.getPrototypeOf(instance);
+            const protos = Object.getPrototypeOf(instance);
+            if (protos) {
+                keys = Object.getOwnPropertyNames(protos);
+            }
         }
-        for (let key in keys) {
-            let member = instance[key];
+        for (const key of keys) {
+            const member = instance[key];
             if (key !== "constructor" && typeof member === "function") {
                 Binder.bind(instance, key, member);
             }
         }
+    }
+
+    /**
+     * Constructs Binder class.
+     */
+    public constructor() {
+        Binder.bindAll(this);
     }
 }
