@@ -48,7 +48,7 @@ export interface IType<T> {
      * @param o
      * @return {number}
      */
-    getSize?(o: T): number
+    getSize?(o: T): number;
 
     /**
      *
@@ -57,7 +57,6 @@ export interface IType<T> {
      */
     equals?(src: T, dest: T): boolean;
 }
-
 
 /**
  * When you use Object.prototype.toString.call(src) then it returns somethings like that [object ...].
@@ -74,75 +73,13 @@ const toString = Object.prototype.toString;
 
 export default class Type<T> implements IType<T> {
     /**
-     * Checks the given value is Empty or not.
-     * @param o
-     * @return {boolean}
-     */
-    readonly hasNot: (o: T) => boolean;
-    /**
-     * Checks the given value is primitive or not. Primitive mean you can use `=` (equality) sign on primitive types
-     * @return {boolean}
-     */
-    readonly isPrimitive: () => boolean;
-    /**
-     * Checks the given value is Json Type or not.
-     * @return {boolean}
-     */
-    readonly isJsonType: () => boolean;
-    /**
-     * Checks the given value is Native Type or not. Native types is used in core javascript library.
-     * @return {boolean}
-     */
-    readonly isNativeType: () => boolean;
-    /**
-     * gets clone of the given value.
-     * @param o
-     * @return {any}
-     */
-    readonly getClone: (o: T) => T;
-    /**
-     * gets type name of the given value.
-     * @param o
-     * @return {string}
-     */
-    readonly getName: (o: T) => string;
-    /**
-     * gets size of the given value
-     * @param o
-     * @return {number}
-     */
-    readonly getSize: (o: T) => number;
-    /**
-     * gets size of the given value
-     * @param o
-     * @return {number}
-     */
-    readonly equals: (src: any, dest: any) => boolean;
-
-    /**
-     * sets initial properties to type from the given properties which implements IType<T> interfaces.
-     * Check the given property if property not exist then apply standart properties which are defined as static members.
-     * @param type
-     */
-    public constructor(type: IType<T>) {
-        this.hasNot = type.hasNot ? type.hasNot : Type.hasNot;
-        this.isPrimitive = type.isPrimitive ? type.isPrimitive : Type.isPrimitive;
-        this.isJsonType = type.isJsonType ? type.isJsonType : Type.isJsonType;
-        this.isNativeType = type.isNativeType ? type.isNativeType : Type.isNativeType;
-        this.getClone = type.getClone ? type.getClone : Type.getClone;
-        this.getName = Type.getName;
-        this.getSize = type.getSize ? type.getSize : Type.getSize;
-        this.equals = type.equals ? type.equals : Type.equals;
-    }
-
-    /**
      * gets Raw Name of the given type like [object ...]
      * @param o
      * @returns {any}
      */
     public static getRawName(o: any): string {
         return toString.call(o);
-    };
+    }
 
     /**
      * gets type name of the given value.
@@ -150,11 +87,11 @@ export default class Type<T> implements IType<T> {
      * @return {string}
      */
     public static getName(o: any): string {
-        let typeObjectString = toString.call(o);
-        let startIndex = TYPE_PREFIX.length;
-        let length = typeObjectString.length - startIndex - 1;
+        const typeObjectString = toString.call(o);
+        const startIndex = TYPE_PREFIX.length;
+        const length = typeObjectString.length - startIndex - 1;
         return typeObjectString.substr(startIndex, length);
-    };
+    }
 
     /**
      * Checks the given value is Empty or not.
@@ -162,7 +99,7 @@ export default class Type<T> implements IType<T> {
      * @return {boolean}
      */
     public static hasNot(o: any) {
-        return o === null || o === undefined
+        return o === null || o === undefined;
     }
 
     /**
@@ -196,7 +133,7 @@ export default class Type<T> implements IType<T> {
      */
     public static getClone<E>(o: E): E {
         return o;
-    };
+    }
 
     /**
      * gets size of the given value
@@ -216,5 +153,67 @@ export default class Type<T> implements IType<T> {
      */
     public static equals(src: any, dest: any): boolean {
         return src === dest;
+    }
+
+    /**
+     * Checks the given value is Empty or not.
+     * @param o
+     * @return {boolean}
+     */
+    public readonly hasNot: (o: T) => boolean;
+    /**
+     * Checks the given value is primitive or not. Primitive mean you can use `=` (equality) sign on primitive types
+     * @return {boolean}
+     */
+    public readonly isPrimitive: () => boolean;
+    /**
+     * Checks the given value is Json Type or not.
+     * @return {boolean}
+     */
+    public readonly isJsonType: () => boolean;
+    /**
+     * Checks the given value is Native Type or not. Native types is used in core javascript library.
+     * @return {boolean}
+     */
+    public readonly isNativeType: () => boolean;
+    /**
+     * gets clone of the given value.
+     * @param o
+     * @return {any}
+     */
+    public readonly getClone: (o: T) => T;
+    /**
+     * gets type name of the given value.
+     * @param o
+     * @return {string}
+     */
+    public readonly getName: (o: T) => string;
+    /**
+     * gets size of the given value
+     * @param o
+     * @return {number}
+     */
+    public readonly getSize: (o: T) => number;
+    /**
+     * gets size of the given value
+     * @param o
+     * @return {number}
+     */
+    public readonly equals: (src: any, dest: any) => boolean;
+
+    /**
+     * sets initial properties to type from the given properties which implements IType<T> interfaces.
+     * Check the given property if property not exist then apply standart properties which are defined as static members.
+     * @param type
+     */
+    public constructor(type: IType<T>) {
+        this.hasNot = type.hasNot ? type.hasNot : Type.hasNot;
+        this.isPrimitive = type.isPrimitive ? type.isPrimitive : Type.isPrimitive;
+        this.isJsonType = type.isJsonType ? type.isJsonType : Type.isJsonType;
+        this.isNativeType = type.isNativeType ? type.isNativeType : Type.isNativeType;
+        this.getClone = type.getClone ? type.getClone : Type.getClone;
+        this.getName = Type.getName;
+        this.getSize = type.getSize ? type.getSize : Type.getSize;
+        this.equals = type.equals ? type.equals : Type.equals;
     }
 }
